run asynchrone and some refactoring
